feat(auth): implemented delete by using auth or by user jwt

+++ TodoApp/Routers/todos.py
@@ -50,9 +50,12 @@ async def add_todo(db: database_injection,user: user_dependency,todo_request: To
     
 @router.put("/todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
 async def update_todo(db: database_injection,
+                      user: user_dependency,
                       todo_request: TodoRequest,
                       todo_id: int= Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=404,detail= 'Not Authenticated')
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id== user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404,detail= 'Item not found for id {todo_id}!')
     else:
@@ -65,10 +68,13 @@ async def update_todo(db: database_injection,
         db.commit()
         
 @router.delete("/todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: database_injection, todo_id:int = Path(gt=0)):
-            todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
-            if todo_model is None:
-                raise HTTPException(status_code=404,detail='Todo not found!')
-            else:
-                db.query(Todos).filter(Todos.id == todo_id).delete()
-                db.commit()
+async def delete_todo(user: user_dependency,db: database_injection, todo_id:int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401,detail='Authentication Failed!')
+    else:
+        todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id== user.get('id')).first()
+        if todo_model is None:
+            raise HTTPException(status_code=404,detail='Todo not found!')
+        else:
+            db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id== user.get('id')).delete()
+            db.commit()
